refactor(ui): split filter-input-with-suggestions into types/utils files
Match the layout used by sibling vector-set components (e.g.
`similarity-search-form`, `command-preview`): types in
`*.types.ts`, helpers in `*.utils.ts` with their own
`*.utils.spec.ts`, and the component file kept to the React
render. Use the shared `uiSrc/constants/keys` for keyboard key
strings instead of redeclaring them locally.

Also trim doc comments back to one-line essentials in the slice
interfaces, the `attributeKeys` utility, and the component itself.
