+++ redisinsight/ui/src/pages/browser/modules/key-details/components/vector-set-details/similarity-search-form/filter-input-with-suggestions/FilterInputWithSuggestions.spec.tsx
@@ -2,11 +2,7 @@ import React from 'react'
 import { fireEvent, screen } from '@testing-library/react'
 import { render } from 'uiSrc/utils/test-utils'
 
-import {
-  FilterInputWithSuggestions,
-  findActiveDotToken,
-  findUsedAttributeKeys,
-} from './FilterInputWithSuggestions'
+import { FilterInputWithSuggestions } from './FilterInputWithSuggestions'
 
 const TEST_ID = 'filter'
 
@@ -28,64 +24,11 @@ const renderForm = (
   return { ...utils, onChange }
 }
 
-describe('findActiveDotToken', () => {
-  it('returns null when caret is at 0', () => {
-    expect(findActiveDotToken('.price', 0)).toBeNull()
-  })
-
-  it('detects a token right after a leading dot', () => {
-    expect(findActiveDotToken('.pri', 4)).toEqual({
-      dotIndex: 0,
-      prefix: 'pri',
-    })
-  })
-
-  it('detects an empty prefix when caret is right after the dot', () => {
-    expect(findActiveDotToken('.', 1)).toEqual({ dotIndex: 0, prefix: '' })
-  })
-
-  it('detects a token after whitespace', () => {
-    expect(findActiveDotToken('a == 1 and .pr', 14)).toEqual({
-      dotIndex: 11,
-      prefix: 'pr',
-    })
-  })
-
-  it('returns null when the dot is preceded by a word char (decimal numbers)', () => {
-    expect(findActiveDotToken('3.14', 4)).toBeNull()
-  })
-
-  it('returns null when there is no dot before caret', () => {
-    expect(findActiveDotToken('price > 5', 9)).toBeNull()
-  })
-
-  it('returns null when a non-word char sits between the dot and caret', () => {
-    expect(findActiveDotToken('.price > 5', 10)).toBeNull()
-  })
-})
-
-describe('findUsedAttributeKeys', () => {
-  it('returns an empty set when there are no dot tokens', () => {
-    expect(findUsedAttributeKeys('price > 5')).toEqual(new Set())
-  })
-
-  it('collects every fully-typed .attribute token', () => {
-    expect(
-      findUsedAttributeKeys('.price > 5 and .category == "books"'),
-    ).toEqual(new Set(['price', 'category']))
-  })
-
-  it('skips decimal numbers (.dot preceded by a word char)', () => {
-    expect(findUsedAttributeKeys('rating > 3.14')).toEqual(new Set())
-  })
-
-  it('excludes the token at excludeDotIndex (the one being typed)', () => {
-    // ".price > 5 and .pr" — active token starts at the second dot (index 15)
-    expect(findUsedAttributeKeys('.price > 5 and .pr', 15)).toEqual(
-      new Set(['price']),
-    )
-  })
-})
+const focusAt = (input: HTMLInputElement, caret: number, lastKey: string) => {
+  fireEvent.focus(input)
+  input.setSelectionRange(caret, caret)
+  fireEvent.keyUp(input, { key: lastKey })
+}
 
 describe('FilterInputWithSuggestions', () => {
   it('does not render the dropdown until the user focuses and types a dot', () => {
@@ -97,10 +40,7 @@ describe('FilterInputWithSuggestions', () => {
 
   it('renders the dropdown filtered by prefix when the user types a dot', () => {
     renderForm({ value: '.c' })
-    const input = screen.getByTestId(TEST_ID) as HTMLInputElement
-    fireEvent.focus(input)
-    input.setSelectionRange(2, 2)
-    fireEvent.keyUp(input, { key: 'c' })
+    focusAt(screen.getByTestId(TEST_ID) as HTMLInputElement, 2, 'c')
 
     expect(screen.getByTestId(`${TEST_ID}-suggestions`)).toBeInTheDocument()
     expect(
@@ -116,46 +56,29 @@ describe('FilterInputWithSuggestions', () => {
 
   it('inserts a suggestion on click and emits onChange', () => {
     const { onChange } = renderForm({ value: '.c' })
-    const input = screen.getByTestId(TEST_ID) as HTMLInputElement
-    fireEvent.focus(input)
-    input.setSelectionRange(2, 2)
-    fireEvent.keyUp(input, { key: 'c' })
+    focusAt(screen.getByTestId(TEST_ID) as HTMLInputElement, 2, 'c')
 
     fireEvent.mouseDown(screen.getByTestId(`${TEST_ID}-suggestion-category`))
     expect(onChange).toHaveBeenCalledWith('.category')
   })
 
   it('omits attributes already used elsewhere in the expression', () => {
-    // `.price > 5 and .` — caret at the trailing dot, `price` already used
     const value = '.price > 5 and .'
     renderForm({ value })
-    const input = screen.getByTestId(TEST_ID) as HTMLInputElement
-    fireEvent.focus(input)
-    input.setSelectionRange(value.length, value.length)
-    fireEvent.keyUp(input, { key: '.' })
+    focusAt(screen.getByTestId(TEST_ID) as HTMLInputElement, value.length, '.')
 
-    expect(screen.getByTestId(`${TEST_ID}-suggestions`)).toBeInTheDocument()
-    expect(
-      screen.getByTestId(`${TEST_ID}-suggestion-category`),
-    ).toBeInTheDocument()
-    expect(
-      screen.getByTestId(`${TEST_ID}-suggestion-color`),
-    ).toBeInTheDocument()
     expect(
       screen.queryByTestId(`${TEST_ID}-suggestion-price`),
     ).not.toBeInTheDocument()
+    expect(
+      screen.getByTestId(`${TEST_ID}-suggestion-category`),
+    ).toBeInTheDocument()
   })
 
   it('does not hide an attribute while the user is editing it in place', () => {
-    // `.pri` — caret inside the only token. The token at dotIndex 0 is the
-    // active one, so it should not be counted as "used" and `price` must
-    // remain visible while the user finishes typing.
     const value = '.pri'
     renderForm({ value })
-    const input = screen.getByTestId(TEST_ID) as HTMLInputElement
-    fireEvent.focus(input)
-    input.setSelectionRange(value.length, value.length)
-    fireEvent.keyUp(input, { key: 'i' })
+    focusAt(screen.getByTestId(TEST_ID) as HTMLInputElement, value.length, 'i')
 
     expect(
       screen.getByTestId(`${TEST_ID}-suggestion-price`),
@@ -163,15 +86,8 @@ describe('FilterInputWithSuggestions', () => {
   })
 
   it('replaces the whole word-token when the caret is mid-token', () => {
-    // `.ca|t` — caret at index 2, the `t` after the caret is still part of
-    // the token. Picking `category` must produce `.category`, not
-    // `.categoryt` (regression test for the half-token replacement bug).
-    const value = '.cat'
-    const { onChange } = renderForm({ value })
-    const input = screen.getByTestId(TEST_ID) as HTMLInputElement
-    fireEvent.focus(input)
-    input.setSelectionRange(2, 2)
-    fireEvent.keyUp(input, { key: 'ArrowLeft' })
+    const { onChange } = renderForm({ value: '.cat' })
+    focusAt(screen.getByTestId(TEST_ID) as HTMLInputElement, 2, 'ArrowLeft')
 
     fireEvent.mouseDown(screen.getByTestId(`${TEST_ID}-suggestion-category`))
     expect(onChange).toHaveBeenCalledWith('.category')
@@ -181,23 +97,18 @@ describe('FilterInputWithSuggestions', () => {
     const openDropdown = () => {
       const utils = renderForm({ value: '.' })
       const input = screen.getByTestId(TEST_ID) as HTMLInputElement
-      fireEvent.focus(input)
-      input.setSelectionRange(1, 1)
-      fireEvent.keyUp(input, { key: '.' })
+      focusAt(input, 1, '.')
       return { ...utils, input }
     }
 
     it('ArrowDown wraps through the suggestion list', () => {
       const { input } = openDropdown()
-      // 3 suggestions: price, category, color (order = order of `suggestions`
-      // prop). activeIndex starts at 0.
       fireEvent.keyDown(input, { key: 'ArrowDown' })
       expect(
         screen.getByTestId(`${TEST_ID}-suggestion-category`),
       ).toHaveAttribute('aria-selected', 'true')
       fireEvent.keyDown(input, { key: 'ArrowDown' })
       fireEvent.keyDown(input, { key: 'ArrowDown' })
-      // wrapped back to index 0
       expect(screen.getByTestId(`${TEST_ID}-suggestion-price`)).toHaveAttribute(
         'aria-selected',
         'true',
@@ -215,12 +126,12 @@ describe('FilterInputWithSuggestions', () => {
 
     it('Enter picks the highlighted suggestion', () => {
       const { input, onChange } = openDropdown()
-      fireEvent.keyDown(input, { key: 'ArrowDown' }) // category
+      fireEvent.keyDown(input, { key: 'ArrowDown' })
       fireEvent.keyDown(input, { key: 'Enter' })
       expect(onChange).toHaveBeenCalledWith('.category')
     })
 
-    it('Tab picks the highlighted suggestion (and preventDefault keeps focus)', () => {
+    it('Tab picks the highlighted suggestion', () => {
       const { input, onChange } = openDropdown()
       fireEvent.keyDown(input, { key: 'Tab' })
       expect(onChange).toHaveBeenCalledWith('.price')
@@ -228,7 +139,6 @@ describe('FilterInputWithSuggestions', () => {
 
     it('Escape closes the dropdown', () => {
       const { input } = openDropdown()
-      expect(screen.getByTestId(`${TEST_ID}-suggestions`)).toBeInTheDocument()
       fireEvent.keyDown(input, { key: 'Escape' })
       expect(
         screen.queryByTestId(`${TEST_ID}-suggestions`),
@@ -240,9 +150,7 @@ describe('FilterInputWithSuggestions', () => {
     renderForm({ value: '.' })
     const input = screen.getByTestId(TEST_ID) as HTMLInputElement
     const wrapper = screen.getByTestId(`${TEST_ID}-wrapper`)
-    fireEvent.focus(input)
-    input.setSelectionRange(1, 1)
-    fireEvent.keyUp(input, { key: '.' })
+    focusAt(input, 1, '.')
 
     expect(wrapper).toHaveAttribute('role', 'combobox')
     expect(wrapper).toHaveAttribute('aria-expanded', 'true')
