+++ redisinsight/ui/src/pages/browser/modules/key-details/components/vector-set-details/similarity-search-form/filter-input-with-suggestions/FilterInputWithSuggestions.types.ts
@@ -1,0 +1,13 @@
+export interface FilterInputWithSuggestionsProps {
+  value: string
+  onChange: (value: string) => void
+  suggestions: string[]
+  placeholder?: string
+  disabled?: boolean
+  'data-testid'?: string
+}
+
+export interface ActiveDotToken {
+  dotIndex: number
+  prefix: string
+}
