+++ redisinsight/ui/src/pages/browser/modules/key-details/components/vector-set-details/similarity-search-form/filter-input-with-suggestions/index.ts
@@ -1,1 +1,5 @@
-export * from './FilterInputWithSuggestions'
+export { FilterInputWithSuggestions } from './FilterInputWithSuggestions'
+export type {
+  ActiveDotToken,
+  FilterInputWithSuggestionsProps,
+} from './FilterInputWithSuggestions.types'
