+++ redisinsight/ui/src/utils/vectorSet/attributeKeys.ts
@@ -1,8 +1,3 @@
-/**
- * Parse a vector set element's `attributes` JSON string and return its
- * top-level keys. Returns `[]` for any non-JSON, non-object, or array value
- * so callers can union safely without guarding for malformed payloads.
- */
 export const extractAttributeKeys = (
   raw: string | undefined | null,
 ): string[] => {
@@ -17,12 +12,8 @@ export const extractAttributeKeys = (
   }
 }
 
-/**
- * Union the alphabetically-sorted top-level attribute keys across `existing`
- * and the keys discovered in `items`. Stable order keeps memoized consumers
- * (e.g. the similarity-search filter dropdown) referentially stable when no
- * new keys appear.
- */
+// Returns the same reference when no new keys are added so memoized consumers
+// stay stable across paginated loads.
 export const mergeAttributeKeys = (
   existing: string[],
   items: { attributes?: string }[],
